+++ metadata_mapper/lambda_function.py
@@ -1,22 +1,24 @@
 import json
-import os
 import sys
 from urllib.parse import urlparse, parse_qs
 import settings
 import importlib
-from mapper import UCLDCWriter, Record, VernacularReader
+from mappers.mapper import UCLDCWriter, Record, VernacularReader
 
 
 def import_vernacular_reader(mapper_type):
+    """
+    accept underscored_module_name_prefixes
+    accept CamelCase class name prefixes split on underscores
+    for example:
+    mapper_type | mapper module       | class name
+    ------------|---------------------|------------------
+    nuxeo       | nuxeo_mapper        | NuxeoVernacular
+    content_dm  | content_dm_mapper   | ContentDmVernacular
+    """
     mapper_module = importlib.import_module(
-        f"{mapper_type}_mapper", package="metadata_mapper")
+        f"mappers.{mapper_type}_mapper", package="metadata_mapper")
 
-    # accept CamelCase class name conventions split on underscores,
-    # for example:
-    # mapper_type | mapper module       | class name
-    # ------------|---------------------|------------------
-    # nuxeo       | nuxeo_mapper        | NuxeoVernacular
-    # content_dm  | content_dm_mapper   | ContentDmVernacular
     mapper_type_words = mapper_type.split('_')
     class_type = ''.join([word.capitalize() for word in mapper_type_words])
     vernacular_class = getattr(
@@ -35,7 +37,10 @@ def parse_enrichment_url(enrichment_url):
                        .replace('-', '_'))
     kwargs = parse_qs(urlparse(enrichment_url).query)
     if enrichment_func not in dir(Record):
-        raise Exception(f"ERROR: {enrichment_func} not found in {Record}")
+        if settings.SKIP_UNDEFINED_ENRICHMENTS:
+            return None, None
+        else:
+            raise Exception(f"ERROR: {enrichment_func} not found in {Record}")
     return enrichment_func, kwargs
 
 
@@ -54,6 +59,8 @@ def map_page(payload, context):
 
     for enrichment_url in collection.get('rikolti__pre_mapping'):
         enrichment_func, kwargs = parse_enrichment_url(enrichment_url)
+        if not enrichment_func and settings.SKIP_UNDEFINED_ENRICHMENTS:
+            continue
         print(f"running enrichment: {enrichment_func} with {kwargs}")
         source_metadata_records = [
             record.enrich(enrichment_func, **kwargs)
@@ -68,6 +75,8 @@ def map_page(payload, context):
 
     for enrichment_url in collection.get('rikolti__enrichments'):
         enrichment_func, kwargs = parse_enrichment_url(enrichment_url)
+        if not enrichment_func and settings.SKIP_UNDEFINED_ENRICHMENTS:
+            continue
         if enrichment_func in ['required_values_from_collection_registry',
                                'set_ucldc_dataprovider']:
             kwargs.update({'collection': collection})
