+++ metadata_mapper/mappers/nuxeo_mapper.py
@@ -1,5 +1,5 @@
 import json
-from mapper import VernacularReader, Record
+from .mapper import VernacularReader, Record
 
 
 class NuxeoRecord(Record):
