+++ metadata_mapper/settings.py
@@ -3,6 +3,7 @@ import os
 LOCAL_RUN = os.environ.get('MAPPER_LOCAL_RUN', False)
 DATA_SRC = os.environ.get('MAPPER_DATA_SRC', 's3')
 DATA_DEST = os.environ.get('MAPPER_DATA_DEST', 's3')
+SKIP_UNDEFINED_ENRICHMENTS = os.environ.get('SKIP_UNDEFINED_ENRICHMENTS', False)
 
 LOG_LEVEL = os.environ.get('LOG_LEVEL', 'DEBUG')    # doesn't currently do anything
 S3_BUCKET = os.environ.get('S3_BUCKET', False)
